lib/python/opentoken-pyspark/requirements.txt
# PySpark for distributed processing
pyspark==3.5.6

# Cryptography for token decryption in overlap analysis
pycryptodome>=3.18.0

# OpenToken core library (local install required)
# Install with: pip install -e ../opentoken
